requirements.txt
pillow==11.2.1
psycopg2-binary==2.9.10